requirements.txt
torch>=1.13.0
transformers>=4.30.0
mlflow>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
tensorflow
pandas 
numpy
scikit-learn
fastapi
uvicorn
nltk
rouge_score
sentencepiece
selenium
webdriver-manager
xgboost
statsmodels
prophet
fastapi
uvicorn[standard]
prometheus-client
dvc
dagshub
-e .